refactor: DB 연결을 위해 Repository 구현체 변경
- ArticleRepository를 인터페이스로 구현
- 기존의 ArticleRepository는 MemoryArticleRepository로 변경 후,
  @Repository 어노테이션 제거
- ArticleService에 구현체가 아닌 상위 타입을 필드로 설정
