+++ src/main/java/codesquad/springcafe/service/ArticleService.java
@@ -30,6 +30,6 @@ public class ArticleService {
     }
 
     public List<Article> getArticles() {
-        return articleRepository.getArticles();
+        return articleRepository.findAll();
     }
 }
